refactor: PickTagPicker 이름 변경
