+++ frontend/techpick/src/components/PickTagPicker/index.tsx
@@ -1,3 +1,3 @@
-export { TagPicker } from './TagPicker';
+export { PickTagPicker } from './PickTagPicker';
 export { DeleteTagDialog } from './DeleteTagDialog';
 export { DeselectTagButton } from './DeselectTagButton';
